Fix issue #245 part 2)
Perform a better overflow check on integer values, especially in the case
of negative values. If we have a Token::TOKEN_OP_NEG in front of a
Token::TOKEN_INT_LITERAL, the parser will negate the value of
Token::TOKEN_INT_LITERAL, then perform the parsing of the literal as
usual. This reduces the number of code paths checking for integer bounds.

TODO:
- Constants still escape bound checking in many cases, but this is more
  hard to take into account. The cleanest implementation would use a
  preprocessor to replace constants' occurences, before performing
  the regular parser pass.