Fix issue #245 part 1)
0xFFFF should be equal to -1, but the current lexer returns
65535. This value is then converted to -1 only when generating the
bytecode. Thus, any bound checking becomes really hard.

This patch fixes the lexer to correctly return a signed value for binary
and hexadecimal values. On the other hand, do not convert integer value
to the signed equivalent, as the user is expected to enter a signed
value. Thus, values above 32767 must be illegal.