[issue 268] Detect the case (fixup will come later)
Add a helper function to detect if a MemoryVector with a given name
exists inside the tree. If the lhs appears also as a rhs inside an
AssignmentNode, a temporary variable should be added during
expandVectorialNodes():

foo = something with foo

should become

temp = something with foo
foo = temp

To add the temporary variable, we still need some more work, so for now
only detect this case.