[issue 268] Pass a pointer to the current compiler during expandVectorialNodes()
If we need to create a temporary variable, we need a way to access the
current compiler object, as it contains the memory maps and supporting
functions. The simplest solution is to pass Compiler* as an argument of
expandVectorialNodes().

NB: this appeals for a better mechanism. The memory managment could be
part of a dedicated class, accessed both by Compiler and Node classes.
This would be easier to manage, but we still would need a pointer to it.