add a verbose option to the "load aseba code" interface
By default, the service returns 200: ok even if the compilation
failed. This is problematic if the caller needs to be aware of any
compilation error. To fix that, add a verbose option to the content
of PUT /nodes/:NODENAME.

Example:
    PUT with file=<code> behaves normally
    PUT with verbose=true,file=<code> returns an empty body upon
    success and the error message (utf string) upon failure

Note that the verbose option MUST BE FIRST or it will be interpreted
as part of the code to send.