add reset/stop node + reset_all/stop_all functionalities
By default, /reset will stop the motors and rerun the code. But sometimes
we want to stop the code once and for all. Also, what if we have
multiple targets ? Can't we reset just one ?

With this commit, we propose the following:

   GET /reset or /reset_all : stop and rerun all nodes
   GET /stop or /stop_all : stop all nodes, no rerun
   GET /nodes/:NAME/!reset : stop and rerun node :NAME
   GET /nodes/:NAME/!stop : stop node :NAME, no rerun

Notice the "!" before the keywords. Without it, the request will be
interpreted as a query for a variable named "reset" or "stop".