Add a periodic event block, variant 2.
This gives access to the Thymio's second timer and makes things like blinking LEDs (the "Hello World" of microcontroller programming) much easier. A limited form of periodic events has previously been available by accident from button and proximity event blocks with no buttons/sensors enabled, but this addition provides a discoverable user interface for such functionality and makes it configurable. Events can be triggered on opposite phases of the period (either on the "tick" or on the "tock" of the pendulum), further simplifying alternating actions like blinking.