Source tree reorganization.
This clean up the layout of the project.

Sources are moved to aseba/
Some docs are move to docs/. This will be improved later.

Compilation of examples and tests is intentionnally broken and
will be fixed in a subsequent commit.

Maintainers scripts have been modified to been callable
from anywhere.