Improve version number handling
The file VERSION.txt describes the version of Aseba.
CMake uses that file to construct the version number of
the package.

That version is also embedded in the `asebacommon` library by
CMake.

`common/consts.h` defines two variables:

* ASEBA_VERSION  : A textual representation of aseba version ("x.y.z")
* ASEBA_REVISION : A textual representation of the git short hash

When releasing a new version of the software, only VERSION.txt
needs to be updated.